+++ cx-portal/src/components/pages/AppReleaseProcess/components/AppPage/index.tsx
@@ -210,20 +210,24 @@ export default function AppPage() {
           <InputLabel sx={{ mb: 3, mt: 3 }}>
             {t('content.apprelease.appPage.images') + ' *'}
           </InputLabel>
-          <Controller
-            name={'images'}
-            control={control}
-            rules={{
-              required: true,
+          <ConnectorFormInputField
+            {...{
+              control,
+              trigger,
+              errors,
+              name: 'images',
+              type: 'dropzone',
+              acceptFormat: {
+                'image/png': [],
+                'image/jpeg': [],
+              },
+              maxFilesToUpload: 3,
+              rules: {
+                required: {
+                  value: true,
+                },
+              },
             }}
-            render={({ field: { onChange, value } }) => (
-              <Dropzone
-                onFileDrop={(files: File[]) => {
-                  onChange(files[0].name)
-                  trigger('images')
-                }}
-              />
-            )}
           />
           {errors?.images?.type === 'required' && (
             <p className="file-error-msg">
@@ -257,6 +261,11 @@ export default function AppPage() {
                   errors,
                   name: item,
                   type: 'dropzone',
+                  acceptFormat: {
+                    'image/png': [],
+                    'image/jpeg': [],
+                  },
+                  maxFilesToUpload: 1,
                   rules: {
                     required: {
                       value: true,