+++ RELEASENOTES.md
@@ -33,7 +33,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Bugfix - Text updates on company data overlay
 * Bugfix - Fixed modal width, subscribe refetch and services loading effect
 * Bugfix - User Management - AddUser Roles missing
-
+* Feature - App Release Process - upload
 
 ### 0.5.3